refactored and moved all methods generating different git URLs to the new helper file
