+++ public/src/componentHelper.tsx
@@ -56,3 +56,37 @@ export function getActiveCourses(courses: Course[], enrols: Enrollment[], userID
     });
     return activeCourses;
 }
+
+
+export function groupRepoLink(groupName: string, courseURL: string): JSX.Element {
+    return <a href={courseURL + slugify(groupName)} target="_blank">{ groupName }</a>;
+}
+
+function gitUserLink(user: string): string {
+    return "https://github.com/" + user;
+}
+
+function labRepoLink(course: string, login: string): string {
+    return course + login + "-labs";
+}
+
+export function generateGitLink(user: string, courseURL?: string): string {
+    return courseURL ? labRepoLink(courseURL, user) : gitUserLink(user);
+}
+
+export function slugify(str: string): string {
+
+    str = str.replace(/^\s+|\s+$/g, "").toLowerCase();
+
+    // Remove accents, swap ñ for n, etc
+    const from = "ÁÄÂÀÃÅČÇĆĎÉĚËÈÊẼĔȆÍÌÎÏŇÑÓÖÒÔÕØŘŔŠŤÚŮÜÙÛÝŸŽáäâàãåčçćďéěëèêẽĕȇíìîïňñóöòôõøðřŕšťúůüùûýÿžþÞĐđßÆa·/_,:;";
+    const to   = "AAAAAACCCDEEEEEEEEIIIINNOOOOOORRSTUUUUUYYZaaaaaacccdeeeeeeeeiiiinnooooooorrstuuuuuyyzbBDdBAa------";
+    for (let i = 0 ; i < from.length ; i++) {
+        str = str.replace(new RegExp(from.charAt(i), "g"), to.charAt(i));
+    }
+
+    // Remove invalid chars, replace whitespace by dashes, collapse dashes
+    str = str.replace(/[^a-z0-9 -]/g, "").replace(/\s+/g, "-").replace(/-+/g, "-");
+
+    return str;
+}
