+++ public/src/components/teacher/CourseGroup.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 import { Course, Group, User, Status } from '../../../proto/ag_pb';
 import { BootstrapButton, DynamicTable, Search } from "../../components";
-import { bindFunc, RProp, generateLabRepoLink } from '../../helper';
+import { bindFunc, RProp } from '../../helper';
 import { CourseManager, ILink, NavigationManager } from "../../managers";
 import { BootstrapClass } from "../bootstrap/BootstrapButton";
-import { generateGroupRepoLink } from "./labHelper";
+import { groupRepoLink, generateGitLink } from '../../componentHelper';
 
 interface ICourseGroupProps {
     approvedGroups: Group[];
@@ -103,7 +103,7 @@ export class CourseGroup extends React.Component<ICourseGroupProps, ICourseGroup
 
     private renderRow(group: Group, withLink: boolean): (string | JSX.Element)[] {
         const selector: (string | JSX.Element)[] = [];
-        const groupName = withLink ? generateGroupRepoLink(group.getName(), this.props.courseURL) : group.getName();
+        const groupName = withLink ? groupRepoLink(group.getName(), this.props.courseURL) : group.getName();
         selector.push(groupName, this.getMembers(group.getUsersList()));
         const actionButtonLinks = this.generateGroupButtons(group);
         const actionButtons = this.renderActionRow(group, actionButtonLinks);
@@ -151,7 +151,7 @@ export class CourseGroup extends React.Component<ICourseGroupProps, ICourseGroup
                 separator = " ";
             }
 
-            const nameLink = <span key={"s" + i} ><a href={ generateLabRepoLink(this.props.courseURL, user.getLogin())}
+            const nameLink = <span key={"s" + i} ><a href={ generateGitLink(user.getLogin(), this.props.courseURL)}
              target="_blank">{ user.getName() }</a>{separator}</span>;
             names.push(nameLink);
             });
